requirements.txt
Django==2.1.11
djangorestframework==3.9.1
django-cors-headers==2.4.0
psycopg2==2.7.5
unidecode==1.0.22
Pillow==5.3.0

whitenoise==4.1
Brotli==1.0.6
dj-database-url==0.5.0
raven==6.9.0
django-redis==4.9.0
django-storages==1.7.1
boto3==1.9.14
scout-apm==1.3.7
django-basic-auth-ip-whitelist==0.2.1
pycountry==18.5.26
djangorestframework-jwt==1.11.0
cryptography==2.3.1
requests==2.20.0
mozilla-django-oidc==1.1.2
django-import-export==1.1.0
django-webpack-loader==0.6.0
faker==2.0.0
factory-boy==2.11.1
